requirements.txt
flask
requests
yfinance
python-dotenv
groq
textblob
pandas